requirements.txt
python-telegram-bot[webhooks]
google-genai
SQLAlchemy
psycopg2-binary
pandas
matplotlib